Cleanups, renames, removing do-not-submits

+++ Backend/src/utils/counters.cpp
@@ -36,7 +36,7 @@ LLVMValueRef adjustCounterReturnOld(
     LLVMBuilderRef builder,
     LLVMTypeRef type,
     LLVMValueRef counterPtrLE,
-    int adjustAmount) {
+    int64_t adjustAmount) {
   auto prevValLE = LLVMBuildLoad2(builder, type, counterPtrLE, "counterPrevVal");
   auto adjustByLE = LLVMConstInt(type, adjustAmount, true);
   assert(LLVMTypeOf(prevValLE) == LLVMTypeOf(adjustByLE));
